ENH: Add app watchdog
The app watchdog works by checking if the file /data/micronet_app_watchdog.cfg. If the file is present we set the app watchdog time (value in seconds). If the file is not present or if the value is less than 60sec and greater than 5000sec it disables the app watchdog.

To pet the app watchdog an mctl message 'mctl app_ping' needs to be sent. If the app watchdog is not pet, the periodic MCU to iodriver pings will be stopped. This would end up resetting the MCU and A8 board eventually

Testing: Verified that the pings work by sending them manually(through the console) and the app_watchdog time is read correctly and if not present or invalid, it is disabled.