BUG: Sometimes the MCTL app does not shutdown the device on a reboot
Sometimes the MCTL app does not shutdown the device because
setSysPropPowerCtlShutdown() does not run. This may be because the
activity gets paused. Added an onPause call hoping this fixes the issue.

Further testing shows that it does not fix the issue but may still be a
good thing to have.