Maint: Version 2.2
Fix for orbcomm
This version of the service enables cellular data based on a
configuration file that can be modified by the end user/customer. This
file is used by the service to make a decision of enabling cell data
when the cores are below a certain range (80) irrespective of the
settings. This was introduced to add flexibility to the service so that
it could prevent non-comm devices.