Fix: Cellular Data Service v2.2.4 (Orbcomm)
On Orbcomms request of always on cell data, Micronet modified the
existing service. The service now reads a configuration file in
/sdcard/MicronetService/CellularTemperatureService before it
enables/disables data.
Summary
- Post interval changed from 10/12s to 30s
- The default setting for override configuration has been set to True.
The service by default will always enable cell data. This change was
introduced because it was found that on an OBC5, customers cannot push
files via Redbend.
- Modified file names, log messages and log files.
- Version name incremented